refactors imports from redux to be from index, adds comments in recipe.js reducer file
